refactor(core): extract OIDC stack to agent-id-core/lib/oidc.mjs
Move the entire OIDC block out of skills/alien-agent-id/lib.mjs:
discovery, authorize/poll/exchange, refresh, userinfo, JWKS fetch,
id_token verification (verifyIdToken and verifyIdTokenSignatureOnly),
plus the supporting internal helpers (TLS-only base URL guard, DPoP
nonce handling per RFC 9449 §8/§9, JSON HTTP wrapper, JWT parse).

The dpopNonceCache singleton stays in the module so the per-process
nonce-batching across token + userinfo + refresh calls keeps working
unchanged. Drops the unused normalizeOptionalString helper.

lib.mjs re-exports the OIDC surface via `export *` and pulls parseJwt,
refreshSession, and verifyIdToken back in for internal use by
SignatureEngine (still in lib.mjs, slated for the next move).

No public-surface changes; all 184 tests pass.
